+++ lib/mmdb2_decoder.ex
@@ -46,18 +46,11 @@ defmodule MMDB2Decoder do
   alias MMDB2Decoder.LookupTree
   alias MMDB2Decoder.Metadata
 
-  @type decode_option ::
-          {:double_precision, Float.precision_range()}
-          | {:float_precision, Float.precision_range()}
-          | {:map_keys, :atoms | :atoms! | :strings}
-
-  @type decode_options_map :: %{
+  @type decode_options :: %{
           optional(:double_precision) => nil | Float.precision_range(),
           optional(:float_precision) => nil | Float.precision_range(),
           optional(:map_keys) => nil | :atoms | :atoms! | :strings
         }
-
-  @type decode_options :: decode_options_map | [decode_option]
 
   @type decoded_value :: :cache_container | :end_marker | binary | boolean | list | map | number
   @type lookup_value :: decoded_value | nil
@@ -161,14 +154,8 @@ defmodule MMDB2Decoder do
       }
   """
   @spec lookup_pointer(non_neg_integer, binary, decode_options) :: {:ok, lookup_value}
-  def lookup_pointer(pointer, data, options \\ @default_decode_options)
-
-  def lookup_pointer(pointer, data, options) when is_map(options) do
+  def lookup_pointer(pointer, data, options \\ @default_decode_options) do
     {:ok, Data.value(data, pointer, options)}
-  end
-
-  def lookup_pointer(pointer, data, options) when is_list(options) do
-    lookup_pointer(pointer, data, Map.new(options))
   end
 
   @doc """